improved connection helper script
the script connect.sh can be used to connect to field nodes via the
manager node as an SSH relay as long as a reverse tunnel exists between
the field node and the manager node. The local port on the manager node
listening to the reverse tunnel must be published as part of the DNS
TXT record for the subdomain.