configuration of power settings on field nodes
Hibernation is now enabled if the machine supports it (if it is a
hardware type listed in node-power-policy.yml, if the swap partition is
larger than the RAM capacity, and if gsettings is present).
critically-low battery level results in a hibernation action where
possible, otherwise a graceful shutdown. lid closure does nothing, and
the machines do not sleep when idle. Power settings are changed via:
“dbus-launch --exit-with-session gsettings set
org.gnome.settings-daemon.plugins.power”