add faked dfree script to samba config
the MiSeq instrument will only write to an output directory if its
volume has >100GB. As expected, this causes issues when the volume has
<100GB. As MiSeq runs are typically <=5GB, we can spoof a larger amount
of free space by overriding the samba “dfree command” with a script
that outputs freespace with a 90GB offset. This ensures that the MiSeq
can still write until there is <10GB of free space on the device. This
is dependent on the windows samba client using the freespace numbers
reported by the samba server.