delay before moving files to external drive
The MiSeq expects the run directory to be present during the post-run wash stage, even if sequencing has finished. This can lead to an error if the upload to DNAnexus is fast and the post-upload script moves files before the wash has finished. The wash takes ~20 minutes, so we'll wait for 30 minutes to give it a chance to finish washing before copying files