do not truncate netbios name since we have netbios disabled
do not truncate netbios name since we have netbios disabled; wsdd2 reads the smb.conf netbios name value via testparm, and advertises it to the network. since we have netbios disabled, we can have a >15 char value for the netbios name (i.e. the full system hostname), and have wsdd2 advertise it to the network so the expected hostname shows up in the Network browser